requirements.txt
pyyaml
sqlalchemy
git+https://github.com/Rapptz/discord.py@async
